Restructured to reflect widget lifecycle of other ports.
Fixed styling of ToolBar which turned out to be the real issue.
Anyway, widget creation is two stage, createHandle to create the
scene graph, and registerHandle to add event handlers.

The widget takedown is the same as the other ports by using the
same implementation of release(boolean) method.

Still have NPEs on Workbench startup.